+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New:
 ----------------------
 - Better stock management when making a formula
+- Add ingredient info in formula making  view